ASoC: SOF: Add Sound Open Firmware KControl support
SOF exposes regular ALSA Kcontrols that are defined by topology. This
patch converts the Kcontrol IO to DSP IPC.

The current implementation is not optimal due to limitations of the
existing firmware: for every get/put the core wakes up the DSP and
generates an IPC. Improvements are already implemented but are not
ready yet to be reviewed, so will come as an update.

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>